requirements.txt
streamlit
mediapipe
opencv-python-headless
scikit-learn
joblib
numpy
Pillow